itinerary_service/pyproject.toml
[project]
name = "itinerary_service"
version = "1.0.0"
description = ""
authors = [
    {name = "Daniil Astapenko", email = "xastapen@mendelu.cz"},
]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.26.0",
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "httpx>=0.26.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.1",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.scripts]
dev = "python -m main"
grpc = "python -m grpc_tools.protoc -I../protos --python_out=./src/grpc_output --pyi_out=./src/grpc_output --grpc_python_out=./src/grpc_output ../protos/itinerary.proto"